backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Object Storage
boto3==1.34.34
minio==7.2.3

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication
pyjwt==2.8.0

# Rate Limiting
redis[hiredis]==5.0.1

# Observability
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
locust==2.20.0

# Payments
stripe==8.1.0